Move type to coordinator file
